refactor(AutoGenerateNumsStrategy) : AutoGenerateStrategy에서 generate 대상이 명확하도록 class 명 변경 refactor(AutoGenerateNumsStrategy) : generate 메소드에서 2차원 배열을 사용하지 않고 generateEachLotto 메소드로 각각의 Lotto를 생성하여 이것을 lottoList에 각각 add하도록 변경 refactor(Prize) : 메소드 이름들이 동사로 시작하도록 변경
