+++ src/main/java/lotto/LottoShop.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain.AutoGenerateStrategy;
+import lotto.domain.AutoGenerateNumsStrategy;
 import lotto.domain.LottoMachine;
 import lotto.view.InputView;
 import lotto.view.ResultView;
@@ -10,7 +10,7 @@ import java.util.List;
 public class LottoShop {
     public static void main(String[] args) {
         int money = InputView.inputMoney();
-        LottoMachine lottoMachine = new LottoMachine(money, new AutoGenerateStrategy());
+        LottoMachine lottoMachine = new LottoMachine(money, new AutoGenerateNumsStrategy());
         ResultView.printBuyableLottoNum(lottoMachine);
         ResultView.printGeneratedLottos(lottoMachine);
