+++ src/main/java/lotto/view/InputView.java
@@ -13,6 +13,7 @@ public class InputView {
     private final static String GUIDE_INPUT_MONEY = "구입금액을 입력해 주세요.";
     private final static String NOTI_INVALID_MONEY = "구입금액으로 유효하지 않습니다.";
     private static final String GUIDE_INPUT_WIN_NUMS = "지난 주 당첨 번호를 입력해 주세요.";
+    private static final int NUMS_PER_LOTTO = 6;
 
     public static int inputMoney() {
         Scanner scanner = new Scanner(System.in);
@@ -49,7 +50,7 @@ public class InputView {
 
     private static List<Integer> toWinningNums(String input) {
         String[] nums = input.split(",");
-        List<Integer> winningNums = new ArrayList<>(LottoMachine.NUMS_PER_LOTTO);
+        List<Integer> winningNums = new ArrayList<>(nums.length);
 
         for (int i = 0; i < nums.length; i++) {
             winningNums.add(Integer.parseInt(nums[i].trim()));
@@ -69,7 +70,7 @@ public class InputView {
     }
 
     static boolean isValidNums(String[] nums) {
-        if (nums.length != LottoMachine.NUMS_PER_LOTTO) {
+        if (nums.length != NUMS_PER_LOTTO) {
             return false;
         }
